feat: 完成评论功能API与异常处理体系
- 实现评论发布的后端API接口
- 完善异常枚举，覆盖用户登录、参数校验等场景
- 统一数据库ID字段为bigint，Java类型对应为Long
- 优化错误提示信息，提升用户体验

+++ src/main/java/cn/iayousa/community/dto/QuestionDTO.java
@@ -5,12 +5,12 @@ import lombok.Data;
 //数据传输用类，存储从question表和user表中查询得到的数据信息
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creatorId;
+    private Long creatorId;
     private Long commentCount;
     private Long viewCount;
     private Long likeCount;
